+++ server/player.h
@@ -243,9 +243,9 @@ public:
 	// Process this player during the server loop.
 	void Process(float fElapsedTime);
 	void BroadcastSyncData();
-	void Say(unsigned char * szText, BYTE byteTextLength);
-	void Privmsg(BYTE byteToPlayerID, unsigned char * szText, BYTE byteTextLength);
-	void TeamPrivmsg(unsigned char * szText, BYTE byteTextLength);
+	void Say(unsigned char * szText, size_t byteTextLength);
+	void Privmsg(BYTE byteToPlayerID, unsigned char * szText, size_t byteTextLength);
+	void TeamPrivmsg(unsigned char * szText, size_t byteTextLength);
 	void SetID(BYTE bytePlayerID) { m_bytePlayerID = bytePlayerID; };
 	
 	void StoreOnFootFullSyncData(ONFOOT_SYNC_DATA * pofSync);
